updating ui and update feature

+++ WeatherApp/Models/WeatherData.swift
@@ -23,7 +23,7 @@ struct Current: Codable {
     let pressureMB: Int
     let humidity: Int
     let feelslikeC: Double
-    let visKM: Int
+    let visKM: Double
     let time: String?
